requirements.txt
Flask
pytesseract
Pillow
yake